Extract Python builtin tests.
