+++ tests/test_hask.py
@@ -639,30 +639,6 @@ class TestList(unittest.TestCase):
         self.assertEqual(20, len(L[0, ..., 19]))
 
 
-class TestPython(unittest.TestCase):
-
-    def test_builtins(self):
-        from hask.Python.builtins import callable, cmp, delattr, divmod
-        from hask.Python.builtins import getattr, hasattr, hash
-        from hask.Python.builtins import hex, isinstance, issubclass, len, oct
-        from hask.Python.builtins import repr, setattr, sorted
-
-        class Example(object):
-            a = 1
-
-        self.assertTrue(callable(__+1))
-        self.assertEqual(1, cmp(10) % 9)
-        self.assertEqual(divmod(5)(2), (2, 1))
-
-        with self.assertRaises(te): cmp(1, "a")
-        with self.assertRaises(te): oct(1.0)
-        with self.assertRaises(te): hex(1.0)
-        with self.assertRaises(te): hasattr(list)(len)
-        with self.assertRaises(te): getattr(list)(len)
-        with self.assertRaises(te): setattr(list)(len)
-        with self.assertRaises(te): delattr(list)(len)
-
-
 class Test_README_Examples(unittest.TestCase):
     """Make sure the README examples are all working"""
